deploy: test with Kubernetes 1.21 instead of 1.18
The deployments for 1.18 get removed in favor of testing with 1.21.
For the two versions which get tested pre-merge, testing of both
"production" and "testing" deployments gets restored the way it
was originally intended (more tests on main worker, less on newly
spawned worker).

Kubernetes 1.21 has additional default values that we need to ignore
when validating deployments.