deploy: storage capacity tracking for 1.21, new sidecars
The new external-provisioner v2.2.0 supports storage capacity tracking
with the v1beta1 API in Kubernetes >= 1.21. Our new deployment files
for 1.21 and the operator enable that for 1.21, but not for 1.19 and
1.20

To simplify the sidecar update in the future, the
hack/bump-image-versions.sh gets added which patches all locations
were we define image versions.

Unnecessary variables in pkg/apis/pmemcsi/v1beta1/deployment_types.go
get removed because they were only used there and patching the file is
simpler when there is only one variable per image.